refactor loader by types
